backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
google-genai>=1.0.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
rank-bm25>=0.2.2
PyMuPDF>=1.24.0
langdetect>=1.0.9
reportlab>=4.2.0
python-dotenv>=1.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
pytest>=8.2.0
httpx>=0.27.0
